requirements.txt
streamlit
pandas
google-api-python-client
python-dotenv
streamlit-option-menu
streamlit-agraph
